+++ apps/core-api-spring/src/main/java/com/lifebit/coreapi/dto/ExerciseRecordRequest.java
@@ -11,7 +11,7 @@ public class ExerciseRecordRequest {
 
     private Integer sets;
     private Integer reps;
-    private Float weight;
+    private Double weight;
 
     private Integer durationMinutes;
     private Integer caloriesBurned;
